+++ src/main/java/com/example/online_banking/model/Card.java
@@ -3,6 +3,7 @@ package com.example.online_banking.model;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -21,7 +22,7 @@ public class Card {
     private Long userId;
 
     @Column(name = "current_balance")
-    private Long currentBalance;
+    private BigDecimal currentBalance;
 
     @Column(name = "active_date")
     private Date cartDate;
